feat: :sparkles: Added the database interface and basic hooks
Added the functions the wikiguesser and ranked to make the code cleaner and an enum to make selection easier, added the ability to update users in the database regardless if they are already in there.

a few issues with randomly not sending the follow up exerpt - could be personal wifi or me not focusing the window.